convertclauses method moved to a common place
